refactor(studio): clean up function privileges queries
